+++ backend/core/settings/base.py
@@ -116,6 +116,7 @@ EMAIL_USE_TLS = True
 EMAIL_HOST_USER = os.environ.get("EMAIL_HOST_USER")
 EMAIL_HOST_PASSWORD = os.environ.get("EMAIL_HOST_PASSWORD")
 DEFAULT_FROM_EMAIL = os.environ.get("DEFAULT_FROM_EMAIL")
+FRONTEND_URL = os.environ.get("FRONTEND_URL", "http://127.0.0.1:3000")
 
 LANGUAGE_CODE = "pt-pt"
 TIME_ZONE = "Africa/Luanda"